Trees now on floor

+++ ex5/src/pepse/world/trees/Tree.java
@@ -36,7 +36,7 @@ public class Tree {
                     this.layer);
             trunk.create();
             TreeTop top = new TreeTop(gameObjects, new Vector2(x,
-                    (float)(this.windowDimensions.y()- (trunk.getHeight()+ ground.groundHeightAt(x)))));
+                    (this.windowDimensions.y() - (trunk.getHeight()+ ground.groundHeightAt(x)))));
             top.create();
             }
     }
